refactor: dedupe Retry-After parsing, document 202 skip behavior
Addresses review feedback on #79:

- Extract Retry-After parser into src/http-utils.mjs (parseRetryAfterSeconds
  + getRetryAfterSecondsFromResponse), replacing the duplicated copies in
  index.mjs and pipeline.mjs. The module documents the integer-seconds-only
  assumption (no HTTP-date support; only producer is Blossom).
- Add inline comments at both vttResponse.status === 202 sites in
  pipeline.mjs (classifyVideoOnly and moderateVideo) explaining that pending
  transcripts are a terminal skip for this moderation run and reprocess-on-
  arrival is deliberately deferred to a follow-up (see issue #87).

No behavior change in index.mjs or pipeline.mjs other than the swap to the
shared parser.
